Basic dithering for Linux OpenGL
Add dithering to yuv2rgb shader. This is not optimal since scaling and
possibly other filtering is done later. Dithering should be the final
step.

PBO resolution needs to be made higher to avoid banding from RGB range
expansion in colorspace conversion when two pass rendering is used.

The dither value scaling is a little ugly, but works well. Even with
quantization to 16 levels (4bpc) picture is reasonable on a computer
screen.